sysdash/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
channels>=4.0
channels-redis>=4.1
django-cors-headers>=4.3
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
daphne>=4.0
